requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-jose
pandas
joblib
scikit-learn
xgboost